Huge Update!!!!
So in this update i added the ImageDisplayer class which is a secondary gui opened by the original to allow better viewing of the saved recipes. It is meant to replace the dropdown menu for saved recipes. This code was developed as a standalone gui, before being added to this project and has been fully tested and works. This is still in progress.  I will be updating the documentation on it...